pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "Flask>=3.0.0",
    "firebase-admin>=6.5.0",
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-vision>=3.10.2",
    "numpy>=2.3.4",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]